msm: camera: isp: Fix support of diagnostics module
Support for diagnostics module on VFE side is partially
implemented currently. This change fills the gaps in
support, and streamlines the code to accommodate
expansion of these registers for v1080.

CRs-Fixed: 3856078

+++ drivers/cam_isp/isp_hw_mgr/cam_ife_hw_mgr.h
@@ -102,7 +102,7 @@ struct cam_ife_hw_mgr_debug {
 	uint64_t       csid_debug;
 	uint32_t       rx_capture_debug;
 	uint32_t       enable_recovery;
-	uint32_t       camif_debug;
+	uint64_t       camif_debug;
 	uint32_t       enable_csid_recovery;
 	uint32_t       sfe_debug;
 	uint32_t       sfe_sensor_diag_cfg;
